+++ web/apps/photos/src/components/FixCreationTime.tsx
@@ -2,6 +2,7 @@ import { FocusVisibleButton } from "@/base/components/mui/FocusVisibleButton";
 import type { ModalVisibilityProps } from "@/base/components/utils/modal";
 import log from "@/base/log";
 import { downloadManager } from "@/gallery/services/download";
+import { extractExifDates } from "@/gallery/services/exif";
 import { fileLogID, type EnteFile } from "@/media/file";
 import {
     decryptPublicMagicMetadata,
@@ -11,7 +12,6 @@ import {
 } from "@/media/file-metadata";
 import { FileType } from "@/media/file-type";
 import { PhotoDateTimePicker } from "@/new/photos/components/PhotoDateTimePicker";
-import { extractExifDates } from "@/new/photos/services/exif";
 import {
     Dialog,
     DialogContent,
