+++ web/apps/photos/src/components/PhotoViewer/index.tsx
@@ -12,6 +12,7 @@ import log from "@/base/log";
 import { FileInfo, type FileInfoProps } from "@/gallery/components/FileInfo";
 import { type FileInfoExif } from "@/gallery/components/viewer/data-source";
 import { downloadManager } from "@/gallery/services/download";
+import { extractRawExif, parseExif } from "@/gallery/services/exif";
 import type { Collection } from "@/media/collection";
 import { fileLogID, type EnteFile } from "@/media/file";
 import { FileType } from "@/media/file-type";
@@ -19,7 +20,6 @@ import { isHEICExtension, needsJPEGConversion } from "@/media/formats";
 import { ConfirmDeleteFileDialog } from "@/new/photos/components/FileViewerComponents";
 import { ImageEditorOverlay } from "@/new/photos/components/ImageEditorOverlay";
 import { moveToTrash } from "@/new/photos/services/collection";
-import { extractRawExif, parseExif } from "@/new/photos/services/exif";
 import { usePhotosAppContext } from "@/new/photos/types/context";
 import AlbumOutlinedIcon from "@mui/icons-material/AlbumOutlined";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
